requirements.txt
python-telegram-bot==13.7
requests==2.28.1
beautifulsoup4==4.11.1
aiohttp==3.8.1
python-dotenv==0.20.0
